+++ app/src/main/java/lavender/client/android/data/grpc/GrpcClient.kt
@@ -78,6 +78,8 @@ object GrpcClient {
 
     fun disconnect() = realGrpcClient.disconnect()
 
+    fun getChannel(): io.grpc.ManagedChannel? = realGrpcClient.getChannel()
+
     fun clearLastChatRequestPrefs() = realGrpcClient.clearLastChatRequestPrefs()
 
     fun clearSystemNotification() = realGrpcClient.clearSystemNotification()